All references are moved to the reference plugin. BNB and GRAO are now created on PDF files and will be possible to be printed 2 sided on pages.

+++ system/module/File/php/FileAPP.php
@@ -15,11 +15,13 @@ class FileAPP extends \system\module\File\php\File{
     }
     
     public function delete_dir($path) {
-        $files = glob($path . '/*');
-    	foreach ($files as $file) {
-    		is_dir($file) ? $this->delete_dir($file) : unlink($file);
-    	}
-    	rmdir($path);
+        if(is_dir($path)){
+            $files = glob($path . '/*');
+            foreach ($files as $file) {
+                is_dir($file) ? $this->delete_dir($file) : unlink($file);
+            }
+            rmdir($path);
+        }
     	return;
     }
